requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
PyYAML>=6.0

# PDF processing
pypdf>=3.15.1
pdf2image>=1.16.3
pytesseract>=0.3.10
Pillow>=10.0.0

# LLM integration
openai>=1.0.0
langchain>=0.0.267

# Database
SQLAlchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.1
python-dateutil>=2.8.2